# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Read Station API",
        "description": "API for Read Station",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/suppliers": {
            "get": {
                "tags": [
                    "Admin / Supplier"
                ],
                "summary": "Danh sách nhà cung cấp",
                "description": "Lấy danh sách nhà cung cấp",
                "operationId": "getAllSuppliers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Số lượng mục trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "deleted"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all authors successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/suppliers/{id}": {
            "get": {
                "tags": [
                    "Admin / Supplier"
                ],
                "summary": "Chi tiết nhà cung cấp",
                "description": "Lấy thông tin chi tiết của một nhà cung cấp",
                "operationId": "getSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của nhà cung cấp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get supplier successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/suppliers/create": {
            "post": {
                "tags": [
                    "Admin / Supplier"
                ],
                "summary": "Tạo nhà cung cấp",
                "description": "Tạo mới một nhà cung cấp",
                "operationId": "createSupplier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create supplier successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/suppliers/update/{id}": {
            "put": {
                "tags": [
                    "Admin / Supplier"
                ],
                "summary": "Cập nhật nhà cung cấp",
                "description": "Cập nhật thông tin của một nhà cung cấp",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của nhà cung cấp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update supplier successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/suppliers/delete/{id}": {
            "delete": {
                "tags": [
                    "Admin / Supplier"
                ],
                "summary": "Xóa nhà cung cấp",
                "description": "Xóa một nhà cung cấp",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của nhà cung cấp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete supplier successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}